+++ core/common/src/main/kotlin/com/github/jcornaz/miop/TerminalOperators.kt
@@ -5,6 +5,11 @@ import kotlinx.coroutines.channels.ReceiveChannel
 import kotlinx.coroutines.channels.SendChannel
 import kotlinx.coroutines.channels.consumeEach
 
+/**
+ * Consume this [ReceiveChannel] and send all elements to [target].
+ *
+ * **DO NOT** close [target]
+ */
 @ObsoleteCoroutinesApi
-internal suspend inline fun <T> ReceiveChannel<T>.sendTo(target: SendChannel<T>): Unit =
+public suspend inline fun <T> ReceiveChannel<T>.sendTo(target: SendChannel<T>): Unit =
     consumeEach { target.send(it) }