+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `parallel` operator to easily build parallel pipelines.
 * `parallelMap` and `parallelFilter` operators.
 * `produce(Iterable)` and `produce(Sequence)` functions. They replace `Iterable.openSubscription`, `Sequence.openSubscription` which are deprecated.
+* `ReceiveChannel.sendTo(SendChannel)` suspending extension function which allow to pipe a `ReceiveChannel` to a `SendChannel`.
  
 ### Dependencies
 * Update Kotlin from `1.3.0-rc-57` to `1.3.10`